% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{align_grid_row}
\alias{align_grid_row}
\title{align_grid_row}
\usage{
align_grid_row(plate, cond, side = "t", view = TRUE)
}
\arguments{
\item{plate}{a screenmill plate object loaded by the read_plate function}

\item{cond}{a quoted conditional statement to select range of rows or cols}

\item{side}{one of "t", "b"}

\item{view}{logical display plate with grid defaults to TRUE}
}
\description{
align_grid_row
}
\details{
`align_grid_row` adds a quoted conditional so that grid elements
   can be moved by any combination of colony_row and colony_col.
   This can be used to specify a specific cell. The conditional needs to
   be captured with `quo` as in this example:
        `cond = quo(colony_col%in%c(14:22))`
   The above conditional would affect only columns
   14 to 22 and affect whichever grid element is
   specified in the `t`, or `b` parameters
}
