% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{justify_grid_edge}
\alias{justify_grid_edge}
\title{justify_grid_edge}
\usage{
justify_grid_edge(plate, cond, side = "t", fcn, view = TRUE)
}
\arguments{
\item{plate}{a screenmill plate object loaded by the read_plate function}

\item{cond}{a quoted conditional statement to select range of rows or cols}

\item{side}{one of "t", "b", "l", "r"}

\item{fcn}{function by which to align selected edge, median, max or min suggested}

\item{view}{logical display plate with grid defaults to TRUE}
}
\description{
justify_grid_edge
}
