% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{read_plate}
\alias{read_plate}
\title{read a single screenmill plate and annotation data by plate_id and display}
\usage{
read_plate(plate = "2016-03-02-001-001-003", dir = "data", view = TRUE)
}
\arguments{
\item{plate}{plate id from a screenmill annotation}

\item{dir}{directory containing screenmill data files}

\item{view}{logical display plate with grid defaults to TRUE}
}
\description{
read a single screenmill plate and annotation data by plate_id and display
}
