% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{regrid}
\alias{regrid}
\title{sm_regrid}
\usage{
regrid(plate_obj, grid_rows, grid_cols, replicates)
}
\arguments{
\item{plate}{a screenmill plate object loaded by the read_plate function}
}
\description{
Input: a plate object from the `read_plate` function which contains
annotation, crop and grid data for a specific plate image.
Grid data is replaced by a call to `screenmill:::locate_grid`.
Used to either redo a grid or make a de novo grid after fixing crop issues
}
