% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{save_plate_calibration}
\alias{save_plate_calibration}
\title{save_plate_calibration}
\usage{
save_plate_calibration(plate)
}
\description{
saves modified annotation data to the appropriate screenmill
files after modifications
}
