% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{shift_crop}
\alias{shift_crop}
\title{shift_crop}
\usage{
shift_crop(plate, left = 0, right = 0, up = 0, down = 0, view = TRUE)
}
\arguments{
\item{plate}{a screenmill plate object loaded by the read_plate function}

\item{left}{number of pixels to move grid to the left}

\item{right}{number of pixels to move grid to the right}

\item{up}{number of pixels to move grid up}

\item{down}{number of pixels to move grid down}

\item{view}{logical display plate with grid defaults to TRUE}

\item{cond}{a quoted conditional statement to select range of rows or cols}
}
\description{
shifts the crop of a plate by the specified pixel amount in the specified
direction
}
