% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{shift_grid}
\alias{shift_grid}
\title{shifts plate grid as a unit}
\usage{
shift_grid(plate, left = 0, right = 0, up = 0, down = 0, view = TRUE)
}
\arguments{
\item{plate}{a screenmill plate object loaded by the read_plate function}

\item{left}{number of pixels to move grid to the left}

\item{right}{number of pixels to move grid to the right}

\item{up}{number of pixels to move grid up}

\item{down}{number of pixels to move grid down}

\item{view}{logical display plate with grid defaults to TRUE}
}
\description{
shifts plate grid as a unit
}
