% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{sm_regrid}
\alias{sm_regrid}
\title{sm_regrid}
\usage{
sm_regrid(
  plate_obj,
  grid_rows,
  grid_cols,
  replicates,
  colony_radius = 1,
  max_smooth = 5,
  view = TRUE
)
}
\arguments{
\item{plate_obj}{a screenmill plate object loaded by the read_plate function}

\item{grid_rows}{number of colony rows on plate image}

\item{grid_cols}{number of colony columns on plate image}

\item{replicates}{number of replicates in each plate grid. This is always a square; e.g.
1,4,16}

\item{colony_radius}{defaults to 1}

\item{max_smooth}{defaults to 5}
}
\description{
Input: a plate object from the `read_plate` function which contains
annotation, crop and grid data for a specific plate image.
Grid data is replaced by a call to `screenmill:::locate_grid`.
Used to either redo a grid or make a de novo grid after fixing crop issues
}
