% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{sm_resize_crop}
\alias{sm_resize_crop}
\title{sm_resize_crop}
\usage{
sm_resize_crop(plate, left = 0, right = 0, up = 0, down = 0, view = TRUE)
}
\arguments{
\item{plate}{a screenmill plate object loaded by the read_plate function}

\item{left}{number of pixels to move left grid-edge to the left, use negative values to move it to the right}

\item{right}{number of pixels to move right grid-edge to the right, use negative values to move it to the left}

\item{up}{number of pixels to move top grid-edge up, negative moves it down}

\item{down}{number of pixels to move bottom grid-edge down, negative moves it up}

\item{view}{logical display plate with grid defaults to TRUE}
}
\description{
shifts the crop of a plate by the specified pixel amount in the specified
direction
}
