% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{sm_rotate_crop}
\alias{sm_rotate_crop}
\title{sm_rotate_crop}
\usage{
sm_rotate_crop(plate, new_rotate = 90, view = TRUE)
}
\arguments{
\item{plate}{a screenmill plate object loaded by the read_plate function}

\item{view}{logical display plate with grid defaults to TRUE}

\item{rotate}{number of pixels to move grid to the left}
}
\description{
rotates the crop of a plate by the specified angle. Default rotation from
standard scan template is 90˚. Providing a crop value of < 90˚ gives an
apparent counter-clockwise rotation to the plate. An angle of > 90˚ gives a
clockwise rotation.
}
