% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{standardize_grid}
\alias{standardize_grid}
\title{standardize_grid}
\usage{
standardize_grid(plate, view = TRUE)
}
\arguments{
\item{plate}{plate id from a screenmill annotation}

\item{view}{logical display plate with grid defaults to TRUE}
}
\description{
Throwing some kind of bug
}
