% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{use_calibration_from}
\alias{use_calibration_from}
\title{use_calibration_from}
\usage{
use_calibration_from(acceptor, donor)
}
\arguments{
\item{acceptor}{A screenmill plate image object produced by `read_plate`}

\item{donor}{A screenmill plate image object produced by `read_plate`}
}
\description{
INPUT
   acceptor - a single screenmill plate image with associated metadata produced by read_plate
               'accepts' the donated grid
   donor    - a single screenmill plate image with associated metadata produced by read_plate
               'donates' a grid
}
\details{
progamatically, the acceptor gives the identity elements of the plate (image-file-path, group, etc).
     Identity data is copied onto the donor object with the desired crop and grid data
     The the 'donor' is updated, which reloads the image using the new image path etc.
Copies grid annotation data from the donor onto the acceptor image
}
