% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{view_plate}
\alias{view_plate}
\title{display single palte image with grid overlay}
\usage{
view_plate(plate, grid_color = "blue")
}
\arguments{
\item{plate}{a screenmill plate object loaded by the read_plate function}

\item{grid_color}{eponymous}
}
\description{
display single palte image with grid overlay
}
