% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{align_grid_row}
\alias{align_grid_row}
\title{align_grid_row}
\usage{
align_grid_row(plate, cond, side = "t", view = TRUE)
}
\arguments{
\item{plate}{a screenmill plate object loaded by the read_plate function}

\item{cond}{a quoted conditional statement to select range of rows or cols}

\item{side}{one of "t", "b"}

\item{view}{logical display plate with grid defaults to TRUE}
}
\description{
align_grid_row
}
