% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{shift_grid_cell}
\alias{shift_grid_cell}
\title{shift_grid_cell}
\usage{
shift_grid_cell(
  plate,
  cond,
  left = 0,
  right = 0,
  up = 0,
  down = 0,
  view = TRUE
)
}
\arguments{
\item{plate}{a screenmill plate object loaded by the read_plate function}

\item{cond}{a quoted conditional statement to select range of rows or cols}

\item{left}{number of pixels to move grid to the left}

\item{right}{number of pixels to move grid to the right}

\item{up}{number of pixels to move grid up}

\item{down}{number of pixels to move grid down}

\item{view}{logical display plate with grid defaults to TRUE}
}
\description{

}
