% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmill-grid-addin.R
\name{sm_resize_grid_cell}
\alias{sm_resize_grid_cell}
\title{morph_grid_cell | sm_resize_grid_cell}
\usage{
sm_resize_grid_cell(
  plate,
  cond,
  left = 0,
  right = 0,
  up = 0,
  down = 0,
  view = TRUE
)
}
\arguments{
\item{plate}{a screenmill plate object loaded by the read_plate function}

\item{cond}{a quoted conditional statement to select range of rows or cols}

\item{left}{number of pixels to move grid to the left}

\item{right}{number of pixels to move grid to the right}

\item{up}{number of pixels to move grid up}

\item{down}{number of pixels to move grid down}

\item{view}{logical display plate with grid defaults to TRUE}
}
\description{
Modifies grid by specific edges in order to shrink or expand as necessary
    to capture data or avoid plate anomalies (shadows/reflections/contaminants)
    Uses a conditional as in shift_grid_cell so that specific cells/rows/cols can be
    modified. x and y are recalculated when cells are updated. Positive values for
    left,right,up,down move named grid edges in that direction, negative values reverse this
}
\details{
`sm_resize_grid_cell` adds a quoted conditional so that grid elements
   can be altered by any combination of colony_row and colony_col.
   This can be used to specify a specific cell. The conditional needs to
   be captured with `quo` as in this example:
        `cond = quo(colony_row == 16 & colony_col%in%c(14:22))`
   The above conditional would affect only the grid positions in
   row16 and colonies 14 to 22 and affect whichever grid element is
   specified in the `left`, `right`,`up` or `down` parameters
}
